+++ App.tsx
@@ -3,7 +3,6 @@ import { Provider } from 'react-redux';
 import MainNavigator from './navigation/MainNavigator';
 import store from './store'
 
-
 const App: FC = () => {
 
   return (
@@ -12,7 +11,5 @@ const App: FC = () => {
     </Provider>
   );
 };
-
-
 
 export default App;
