+++ assets/colors/colors.tsx
@@ -1,7 +1,9 @@
 const colors = {
     primary: '#385776',
     white: '#FFFFFF',
-    dark: '#212B36'
+    dark: '#212B36',
+    text: '#56626E',
+    border: '#E4E8EB'
 
 }
