+++ navigation/Routes.ts
@@ -1,0 +1,4 @@
+export enum Routes {
+    HOME = "Home",
+    ADD_CRYPTO = "AddCrypto"
+}
