+++ store/actions/crypto.actions.tsx
@@ -7,17 +7,15 @@ export const READ_DATA = 'READ_DATA'
 export const readData = () => {
     return async (dispatch: Dispatch<Action>) => {
         await AsyncStorage.getItem('@coin').then(data => {
-            if (data !== null) {
+            if (data) {
                 const array = JSON.parse(data);
-                console.log(array);
-                console.log(`Estoy en async storage ${array}`);
-
+                // console.log(`Estoy en async storage ${array}`);
                 dispatch({
                     type: READ_DATA,
                     payload: array
                 })
             } else {
-                console.log('No hay nada en async storage');
+                // console.log('No hay nada en async storage');
             }
         })
